refactor(admin): split home copy editor
Move homepage copy validation and form panels out of home-copy-editor.tsx so the main editor owns draft state and save/reset flow while helpers stay locally scoped.

Verified with pnpm exec prettier --check apps/admin/src/app/home-copy-editor.tsx apps/admin/src/app/home-copy-editor-model.ts apps/admin/src/app/home-copy-editor-panels.tsx; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; git diff --check -- apps/admin/src/app/home-copy-editor.tsx apps/admin/src/app/home-copy-editor-model.ts apps/admin/src/app/home-copy-editor-panels.tsx; pnpm validate; pnpm dlx knip --no-exit-code.
